+++ odoo/local-src/specific_subscription/__manifest__.py
@@ -1,0 +1,19 @@
+# -*- coding: utf-8 -*-
+# Copyright 2017 Camptocamp SA
+# License LGPL-3.0 or later (http://www.gnu.org/licenses/lgpl.html).
+{
+    "name": "RocTool specific subscription",
+    "version": "10.0.1.0.0",
+    "depends": [
+        'sale_contract',
+        'specific_sale',
+    ],
+    "author": "Camptocamp SA",
+    "website": "http://www.camptocamp.com",
+    "license": "LGPL-3",
+    "category": "Sale",
+    "data": [
+        'views/subscription_views.xml',
+    ],
+    'installable': True,
+}